+++ NEWS.md
@@ -1,3 +1,5 @@
+# sortable (development version)
+
 # sortable 0.4.6
 
 ## Upgrade sortable.js