+++ NEWS.md
@@ -1,5 +1,7 @@
 # sortable (development version)
 
+* Add ability to switch the orientation of `rank_list()` items to horizontal. #92
+
 # sortable 0.4.6
 
 ## Upgrade sortable.js